removed unused files, shopbot.py renamed to run.py, CryptoAddressGenerator.py moved to utils
