--- users_count_shop.txt
@@ -1,1 +1,0 @@
-1
